refactor(profile): align Kotlin and native field signedness
Every profile field now mirrors the native storage signedness end to end:
uint8_t/uint32_t slots are UInt and uint64_t slots are ULong on the Kotlin
side, the execution getters return uint32_t, and the wire header magic/version
and route wire values are unsigned. Only the int32_t address fields
(pselect_waiter_shift, mcast_waiter_off) stay signed, and the 64-bit transport
glue keeps Long so it can carry negative int32 values and large uint64 values
through the same slots.

+++ src/core/session/runtime_config.cpp
@@ -49,8 +49,8 @@ int32_t ghostlock::config::RuntimeConfig::apply_profile(const profile::TargetPro
     if (!e) return -1;
     int32_t old_main = main_cpu;
     int32_t old_consumer = consumer_cpu;
-    main_cpu = profile->recommended_main_cpu();
-    consumer_cpu = profile->recommended_consumer_cpu();
+    main_cpu = static_cast<int32_t>(profile->recommended_main_cpu());
+    consumer_cpu = static_cast<int32_t>(profile->recommended_consumer_cpu());
     if (runtime_config_validate_cpus(this) != 0) {
         main_cpu = old_main;
         consumer_cpu = old_consumer;
